refactor(checker): split type-node signature helpers
