+++ crates/tsz-checker/src/types/mod.rs
@@ -26,6 +26,7 @@ mod type_node_merged_value_query;
 mod type_node_property_names;
 mod type_node_query_members;
 mod type_node_resolution;
+mod type_node_signature;
 pub(crate) mod unique_symbol_arena;
 pub(crate) mod utilities;
 pub(crate) mod window_global_this_annotation;
